+++ src/features/job/job-postings-box.tsx
@@ -9,6 +9,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import { JobPostingBlockComponent } from '@/features/job/job-posting-block-component';
 import { useState } from 'react';
 import Typography from '@/components/ui/typography';
+import JobPostingPaginationButton from './job-posting-pagination-button';
 
 type Props = {
   userId: string;
@@ -17,7 +18,7 @@ type Props = {
 export type SortOption = 'latest' | 'oldest' | 'deadline' | 'company' | 'bookmark';
 
 const { META_DATA } = QUERY_KEY;
-const JOB_POSTING_DATA_LIMIT = 15;
+export const JOB_POSTING_DATA_LIMIT = 15;
 
 const JobPostingsBox = ({ userId }: Props) => {
   const queryClient = useQueryClient();
@@ -34,11 +35,6 @@ const JobPostingsBox = ({ userId }: Props) => {
     page,
     limit: JOB_POSTING_DATA_LIMIT,
   });
-
-  const totalPageCount = Math.ceil((data?.totalCount ?? 0) / JOB_POSTING_DATA_LIMIT);
-  const currentGroup = Math.floor((page - 1) / 10);
-  const startPage = currentGroup * 10 + 1;
-  const endPage = Math.min(startPage + 9, totalPageCount);
 
   return (
     <>
@@ -79,29 +75,7 @@ const JobPostingsBox = ({ userId }: Props) => {
             ))}
           </div>
 
-          {totalPageCount > 1 && (
-            <div className='mt-6 flex flex-wrap justify-center gap-2'>
-              {startPage > 1 && (
-                <button onClick={() => setPage(startPage - 1)} className='rounded border px-3 py-1 text-sm'>
-                  &laquo;
-                </button>
-              )}
-              {Array.from({ length: endPage - startPage + 1 }, (_, i) => startPage + i).map((pageNumber) => (
-                <button
-                  key={pageNumber}
-                  onClick={() => setPage(pageNumber)}
-                  className={`rounded border px-3 py-1 text-sm ${pageNumber === page ? 'bg-primary-orange-600 text-white' : 'hover:bg-gray-100'}`}
-                >
-                  {pageNumber}
-                </button>
-              ))}
-              {endPage < totalPageCount && (
-                <button onClick={() => setPage(endPage + 1)} className='rounded border px-3 py-1 text-sm'>
-                  &raquo;
-                </button>
-              )}
-            </div>
-          )}
+          <JobPostingPaginationButton totalCount={data.totalCount} page={page} setPage={setPage} />
         </section>
       )}
     </>
